# BE/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing seasons, rounds, and challenges in the Scout Arena application.",
        "title": "Scout Arena API",
        "contact": {
            "name": "API Support",
            "url": "http://www.scoutarena.com/support"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/challenges/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "challenges"
                ],
                "summary": "Lấy chi tiết thử thách theo vòng đấu",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Challenge ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ChallengeResponse"
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Cập nhật thông tin chi tiết của một thử thách dựa trên ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "challenges"
                ],
                "summary": "Cập nhật thử thách",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Round ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dữ liệu cập nhật",
                        "name": "challenge",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ChallengeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ChallengeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Xóa (soft delete) dựa trên ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "challenges"
                ],
                "summary": "Xóa thử thách",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Challenge ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Challenge deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/rounds": {
            "post": {
                "description": "Tạo vòng đấu mới với tên, hình ảnh và thời gian bắt đầu/kết thúc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rounds"
                ],
                "summary": "Tạo mới một vòng đấu",
                "parameters": [
                    {
                        "description": "Thông tin vòng đấu",
                        "name": "season",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RoundRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.RoundResponse"
                        }
                    },
                    "400": {
                        "description": "Lỗi validation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/rounds/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rounds"
                ],
                "summary": "Lấy chi tiết vòng đấu theo mùa giải",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Round ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.RoundResponse"
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Cập nhật thông tin chi tiết của một vòng đấu dựa trên ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rounds"
                ],
                "summary": "Cập nhật vòng đấu",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Round ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dữ liệu cập nhật",
                        "name": "season",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RoundRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.RoundResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Xóa (soft delete) dựa trên ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rounds"
                ],
                "summary": "Xóa vòng đấu",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Round ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Round deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/seasons": {
            "post": {
                "description": "Tạo mùa giải mới với tên, hình ảnh và thời gian bắt đầu/kết thúc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "seasons"
                ],
                "summary": "Tạo mới một mùa giải",
                "parameters": [
                    {
                        "description": "Thông tin mùa giải",
                        "name": "season",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SeasonRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.SeasonResponse"
                        }
                    },
                    "400": {
                        "description": "Lỗi validation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.ArenaStatus": {
            "type": "string",
            "enum": [
                "DRAFT",
                "INCOMING",
                "ACTIVE",
                "FINISHED"
            ],
            "x-enum-varnames": [
                "ArenaDraft",
                "ArenaIncoming",
                "ArenaActive",
                "ArenaFinished"
            ]
        },
        "model.ChallengeRequest": {
            "type": "object",
            "required": [
                "image_url",
                "name",
                "point",
                "round_id",
                "status"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "point": {
                    "type": "integer"
                },
                "round_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/model.ArenaStatus"
                }
            }
        },
        "model.ChallengeResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "point": {
                    "type": "integer"
                },
                "round_id": {
                    "type": "integer"
                },
                "round_name": {
                    "type": "string"
                },
                "season_id": {
                    "type": "integer"
                },
                "season_name": {
                    "type": "string"
                }
            }
        },
        "model.RoundRequest": {
            "type": "object",
            "required": [
                "image_url",
                "name",
                "occurred_at",
                "season_id",
                "status",
                "total_point"
            ],
            "properties": {
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "occurred_at": {
                    "type": "string"
                },
                "season_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/model.ArenaStatus"
                },
                "total_point": {
                    "type": "integer"
                }
            }
        },
        "model.RoundResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "occurred_at": {
                    "type": "string"
                },
                "season_id": {
                    "type": "integer"
                },
                "season_name": {
                    "type": "string"
                },
                "total_point": {
                    "type": "integer"
                }
            }
        },
        "model.SeasonRequest": {
            "type": "object",
            "required": [
                "ended_at",
                "image_url",
                "name",
                "started_at",
                "status"
            ],
            "properties": {
                "ended_at": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.ArenaStatus"
                }
            }
        },
        "model.SeasonResponse": {
            "type": "object",
            "properties": {
                "ended_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                }
            }
        }
    }
}